Conflicts:
	Classes/LinphoneCoreSettingsStore.m
	Classes/LinphoneManager.m
	linphone.xcodeproj/project.pbxproj
	submodules/build/Makefile
	submodules/build/builder-iphone-os.mk

+++ Classes/Model/ChatModel.m
@@ -103,7 +103,7 @@
     if([self chatId] != nil) {
         [chatId release];
     } 
-    chatId = [[NSNumber alloc] initWithInt:sqlite3_last_insert_rowid(database)];
+    chatId = [[NSNumber alloc] initWithInt:(int)sqlite3_last_insert_rowid(database)];
     sqlite3_finalize(sqlStatement);
 }
